Fix tests and circular dependencies
